Shadow Walker: WIP fix for melee weapon model bug. Code refactor.
