+++ sp/src/game/server/server_episodic.vpc
@@ -53,7 +53,9 @@ $Project "Server (Episodic)"
 			$Folder	"Mod"
 			{
 				$File "mod\npc_lost_soul.cpp"
+				$File "mod\npc_shadow_walker.h"
 				$File "mod\npc_shadow_walker.cpp"
+				$File "mod\weapon_custom_melee.h"
 			}
 			$File	"$SRCDIR\game\shared\episodic\achievements_ep1.cpp"
 			$File	"$SRCDIR\game\shared\episodic\achievements_ep2.cpp"
